requirements.txt
flask
flask-cors
ollama
psutil
pyttsx3
pyaudio
requests
pyinstaller
waitress
vosk
sounddevice
redis
edge-tts
pygame
pyautogui
pdfplumber
duckduckgo-search
dateparser
spotipy
python-dotenv
